gcc-4.6: Fix packaging of native toolchains
The base gcc package is missing liblto_plugin.so.  If we try a simple
hello world compile we get something like:

root@p5020-ds:~# gcc hello.c
gcc: fatal error: -fuse-linker-plugin, but liblto_plugin.so not found
compilation terminated.

We need to include liblto_plugin.so in the base gcc package and not
gcc-dev.

Signed-off-by: Kumar Gala <galak@kernel.crashing.org>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>