hob: more reliable disabling of GPLv3 packages
1. reflect GPLv3's presence in INCOMPATIBLE_LICENSE value in the UI

The hob UI currently only supports GPLv3 as a value for
INCOMPATIBLE_LICENSE but doesn't properly reflect whether the value is
already set. This patch rectifies this.

2. don't stomp over other INCOMPATIBLE_LICENSE values when disabling GPLv3

In case the user has other values set for INCOMPATIBLE_LICENSE we don't
want to overwrite the value, we want to modify it.

Fixes [#1286]

Signed-off-by: Joshua Lock <josh@linux.intel.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>