bitbake.conf: remove PACKAGE_EXTRA_ARCHS_tune-XXX default
Because of the way BitBake handles ??= under certain circumstances, this
default setting ends up stepping all over the real setting from the arch
include file. Since virtually all arch include files or tune files define
a real value for this we shouldn't need to have a default (or it needs to
be done in a different way).

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Acked-by: Kumar Gala <galak@kernel.crashing.org>
Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>