bb/cooker: mark parsed conf files as potential cache invalidators
The include parameter to _parse() tells the parser whether to include the
file in __base_depends, the contents of which have their mtime tested to
invalidate the cache when they have changed.

As I understand it we would want all of the configuration files that are
currently handled with the _parse() method to invalidate the cache to set the
default value of the include parameter to True.

Signed-off-by: Joshua Lock <josh@linux.intel.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>