initrdscripts: fix init-live.sh
Fix bug: [YOCTO #686]
Because the variable "ROOT_IMAGE" in init-live.sh is not correct, it fails to
run when using liveCD. Modify value of "ROOT_IMAGE" to "isolinux/rootfs.img".
If we want to use liveCD, we also need to add some kernel options related to
CDROM support when compling kernel.

Signed-off-by: Jingdong Lu <jingdong.lu@windriver.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>